+++ src/mbnavedit/mbnavedit_util.c
@@ -472,9 +472,9 @@ static XmString StringToXmString(char * str) {
 		return (NULL);
 
 	static char *tagBuf = NULL;
-	static int tagBufLen = 0;
+	static unsigned int tagBufLen = 0;
 	static char *textBuf = NULL;
-	static int textBufLen = 0;
+	static unsigned int textBufLen = 0;
 
 	/*
 	 * For expediencies sake, we'll overallocate this buffer so that
@@ -984,7 +984,7 @@ void BX_MENU_POST(Widget p, XtPointer mw, XEvent *ev, Boolean *dispatch) {
 	Widget m = (Widget)mw;
 	XtGetValues(m, args, argcnt);
 	XButtonEvent *e = (XButtonEvent *)ev;
-	if (e->button != button)
+	if ((int) e->button != button)
 		return;
 	XmMenuPosition(m, e);
 	XtManageChild(m);
@@ -1503,7 +1503,7 @@ LFUNC(atoui, unsigned int, (char *p, unsigned int l, unsigned int *ui_return));
 #endif
 
 static unsigned int atoui(char *p, unsigned int l, unsigned int *ui_return) {
-	int n, i;
+	unsigned int n, i;
 
 	n = 0;
 	for (i = 0; i < l; i++)
@@ -2279,7 +2279,7 @@ static void SetImagePixels(
 	char *dst;
 	int nbytes;
 	unsigned int *iptr;
-	int x, y, i;
+	unsigned int x, y, i;
 
 	iptr = pixelindex;
 	if (image->depth == 1) {
@@ -2339,12 +2339,12 @@ static unsigned long byteorderpixel = MSBFirst << 24;
 #endif
 
 static void SetImagePixels32(
-    XImage *image, unsigned int width, unsigned int height,
-    unsigned int *pixelindex, Pixel *pixels) {
+  XImage *image, unsigned int width, unsigned int height,
+  unsigned int *pixelindex, Pixel *pixels) {
 	unsigned char *addr;
 	unsigned int *paddr;
 	unsigned int *iptr;
-	int x, y;
+	unsigned int x, y;
 
 	iptr = pixelindex;
 #ifndef WORD64
@@ -2386,7 +2386,7 @@ static void SetImagePixels16(
     unsigned int *pixelindex, Pixel *pixels) {
 	unsigned char *addr;
 	unsigned int *iptr;
-	int x, y;
+	unsigned int x, y;
 
 	iptr = pixelindex;
 	if (image->byte_order == MSBFirst)
@@ -2413,7 +2413,7 @@ static void SetImagePixels8(
     XImage *image, unsigned int width, unsigned int height,
     unsigned int *pixelindex, Pixel *pixels) {
 	unsigned int *iptr;
-	int x, y;
+	unsigned int x, y;
 
 	iptr = pixelindex;
 	for (y = 0; y < height; y++)
@@ -2431,7 +2431,7 @@ static void SetImagePixels1(
 	unsigned char bit;
 	int xoff, yoff;
 	unsigned int *iptr;
-	int x, y;
+	unsigned int x, y;
 
 	if (image->byte_order != image->bitmap_bit_order)
 		SetImagePixels(image, width, height, pixelindex, pixels);
